Fix to have one version of Studio page
