+++ frontend/src/console/components/CoursesManage/styles.scss
@@ -3,6 +3,7 @@
 .courses-page {
     // Only put this because the page didn't pick up the right font family.
     font-family: $chivo-font;
+    padding: 5%;
 
     p {
         margin-top: 20px;
@@ -47,5 +48,17 @@
         .fas {
             padding-left: 10px;
         }
+    }
+
+    .course-page-title {
+        font-family: PlayfairDisplay;
+        font-size: 30px;
+        font-weight: bold;
+        font-stretch: normal;
+        font-style: normal;
+        line-height: normal;
+        letter-spacing: normal;
+        color: #1e4c59;
+        text-align: left;
     }
 }
