--- frontend/src/newConsole/components/newCoursesManage/displayMessages.ts
@@ -1,35 +1,0 @@
-const messages = {
-  title: {
-    defaultMessage: 'Courses (Studio)',
-    description: 'The component title'
-  },
-  explanation: {
-    defaultMessage:
-      "Courses are created and managed in Studio. Here's what you need to know:",
-    description: 'Explanation about this page'
-  },
-  instructions_access: {
-    defaultMessage:
-      'Select "manage courses" below to access Studio. You\'ll be navigated to a new website.',
-    description: 'Instructions on accessing Studio'
-  },
-  instructions_credentials: {
-    defaultMessage:
-      'No need to sign up! Log into Studio using the same email address and password you use on OpenCraft.',
-    description: 'Instruction regarding credentials'
-  },
-  button_text: {
-    defaultMessage: 'Manage courses',
-    description: 'Button text'
-  },
-  help_text: {
-    defaultMessage: 'Need help?',
-    description: 'Help text'
-  },
-  help_link: {
-    defaultMessage: 'Visit Support',
-    description: 'Help link'
-  }
-};
-
-export default messages;
