--- frontend/src/newConsole/components/newCoursesManage/index.ts
@@ -1,1 +1,0 @@
-export * from './CoursesManage';
